move SendEmailJob.ts under core/jobs/ directory
